requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
isort==5.5.3
joblib==0.16.0
lazy-object-proxy==1.4.3
libvoikko==4.3
mccabe==0.6.1
nltk==3.5
Pillow==7.2.0
pkg-resources==0.0.0
pylint==2.6.0
pytesseract==0.3.6
regex==2020.7.14
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tqdm==4.49.0
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
